requirements.txt
alembic==1.16.2
APScheduler==3.10.4
backoff==2.2.1
coloredlogs==15.0.1
fastapi==0.115.5
psycopg2-binary==2.9.10
pydantic==2.10.3
python-dotenv==1.1.1
requests==2.32.4
SQLAlchemy==2.0.41
uvicorn==0.34.0
